ft(post-user-registration): send user registration details to api
