+++ package.json
@@ -1,0 +1,45 @@
+{
+  "name": "maintenance-app-client",
+  "version": "1.0.0",
+  "description": "Maintenance tracker front end to consume maintenance tracker api",
+  "main": "app.js",
+  "dependencies": {
+    "babel-cli": "^6.26.0",
+    "babel-core": "^6.26.3",
+    "babel-plugin-transform-object-rest-spread": "^6.26.0",
+    "babel-preset-env": "^1.7.0",
+    "cors": "^2.8.4",
+    "dotenv": "^5.0.1",
+    "eslint": "^4.19.1",
+    "express": "^4.16.3",
+    "nodemon": "^1.17.5"
+  },
+  "devDependencies": {
+    "eslint-config-airbnb": "^16.1.0",
+    "eslint-plugin-import": "^2.12.0",
+    "eslint-plugin-jsx-a11y": "^6.0.3",
+    "eslint-plugin-react": "^7.8.2"
+  },
+  "scripts": {
+    "start": "babel-node app.js",
+    "test": "mocha"
+  },
+  "engines": {
+    "node": "9.11.1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/cwizard2011/maintenace-tracker-app.git"
+  },
+  "keywords": [
+    "Client",
+    "Front",
+    "end"
+  ],
+  "author": "Adeoye Peter Adeola",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/cwizard2011/maintenace-tracker-app/issues"
+  },
+  "homepage": "https://github.com/cwizard2011/maintenace-tracker-app#readme"
+}
