+++ public/js/details.js
